Add basic profile handling
This commit adds basic profiling handling support to the tool. It
defines a yaml parser module that handles getting pieces of information
necessary for the encoding and adds the appropriate calls to run(). Also
the run() function is changed so that it takes a path rather than a hash
so it can use the new profile.pm directly.

This is just a starting point, it will be expanded upon in the future as
we need additional options, etc.