Add a profile option to specify an video encoder preset
This commit adds a new parameter to the video typo in the profile files
for the video encoder preset. This is a magic string specific to the
encoder specified in the profile file that dictates the encoder
settings. It is completely dependent on the encoder you select and there
really isn't any guidance on what it should be, it's 100% up to the user
to specify this correctly.