Fix stdout/stderr capture
The capture lib we used previously to capture the output didn't work as
we had hoped. It waits for the system call to exit and then handles
writing the stdout and stderr to a file. We wanted real time output
written to a file, so this commit cheats and just does it in the
subshell.