Use bash to handle stderr redirection
I can't seem to find a good model for doing stderr redirection in perl.
Maybe one day I'll figure out how to do this natively in perl, but for
now lets just use something I know works. This commit adds a bash
wrapper to calling HandBrakeCLI which will do the redirection and then
the run() function is updated to call that script instead.