Add an option to set the encoder tune in profile
This commit adds a new option to the profile yaml files to set the
video encoder tune to run HandBrakeCLI with. The encoder tune, like the
encoder preset is dependent on the video encoder selected and will tune
some of the encoder options for specific content types you're
transcoding. However, the burden is on the user to know what tunes are
available for the encoder selected. You can use HandBrakeCLI to list the
available tunes for an encoder with "HandBrakeCLI --encoder-tune-list
$encoder".