Add option to specify chapters to use in output
Sometimes discs split episodes by chapter ranges instead of using
titles (I still haven't been able to figure out why anyone does this).
But, in this case it would be useful to be able to split output files
based on chapters. This commit adds a chapter argument to the main yaml
file that will enable specifying chapters to use.