Change default behavior for title option
The title option is optional and if it's not specified then a title is
never passed to the CLI. The intent here was for transcoding already
split files that didn't have distinct titles. (so we avoided having to
pass in a 1 to make things work) This patch addresses this default case
by changing it from a lack of a title on the CLI to using
--main-feature instead.

This will work on single title files because well in the case of 1
title that will always be the "main feature". However, in the
multi-title case this will change the default behavior from being a
failure (due to a lack of title) to using HandBrake's main feature
detection algorithm (a super complex algorithm which just finds the
longest title [1] :p)

[1] https://github.com/HandBrake/HandBrake/blob/a8b4a5afbe95e1f08e71c781ac214eba57b3480b/libhb/dvd.c#L63-L79