Use catfile to manage rename paths
Right now when we attempt to rename files after the transcode we naively
concat the directory string with the filename string. This may or may
not work depending on the paths (and filenames). This commit attempts
to fix this by using the catfiles function which should handle the edge
cases better.